Create useCourtArrangements hook
